requirements.txt
# Core dependencies
numpy>=1.20.0
scipy>=1.7.0
pandas>=1.3.0
matplotlib>=3.4.0
PyYAML>=6.0

# Reinforcement Learning
tensorflow>=2.8.0
gymnasium>=0.28.0
stable-baselines3>=2.0.0

# CFD Simulation
pyvista>=0.32.0  # For 3D visualization
# OpenFOAM integration - install separately
# See: https://openfoam.org/download/
# Optional PyFoam for OpenFOAM processing if needed
# pip install PyFoam

# Optimization and Analysis
scikit-learn>=1.0.0
optuna>=2.10.0  # For hyperparameter optimization

# Visualization
tensorboard>=2.8.0
plotly>=5.5.0
tqdm>=4.62.0  # For progress bars

# Testing and Development
pytest>=6.2.0
black>=22.1.0  # Code formatting
flake8>=4.0.0  # Code linting
isort>=5.10.0  # Import sorting

# Documentation
sphinx>=4.4.0
sphinx-rtd-theme>=1.0.0

# Other Utilities
requests>=2.27.0
joblib>=1.1.0
h5py>=3.6.0  # For model storage